refactor(app): split runtime config tests
Move the large inline runtime_config test module out of tools/runtime_config.rs into tools/runtime_config/tests.rs.

Keep the existing delegate prompt test module separate and leave runtime config behavior unchanged.
